transfer created returned true on App, works on pgAdmin and postman

+++ tenmo-server/src/main/java/com/techelevator/tenmo/dao/AccountDao.java
@@ -22,7 +22,7 @@ public interface AccountDao {
 
     Account getAccountByAccountId(int accountId);
 
-    Account getAccountByUserId(Long userId);
+    Account getAccountByUserId(int userId);
